+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "idbox_generator"
-version = "0.1.0"
+version = "0.2.0"
 description = ""
 authors = []
 readme = "README.md"
